+++ src/data_sources/sql/create_query/create_update_many_query/mod.rs
@@ -37,6 +37,7 @@ impl SqlDataSource {
                 offset,
                 subgraph_config,
                 None,
+                false,
             )?;
 
         query.push_str(" SET ");