+++ src/data_sources/sql/create_query/create_update_one_query/mod.rs
@@ -52,6 +52,7 @@ impl SqlDataSource {
                 offset,
                 subgraph_config,
                 None,
+                false,
             )?;
 
         if let Some(nested_query) = nested_query {